# likecollective-indexer/openapi/schema.yaml
openapi: 3.0.3
info:
  title: LikeNFT Indexer API
  version: 0.1.0
paths:
  /stakings:
    get:
      tags:
        - Staking
      summary: Get stakings
      parameters:
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: The number of items to return.
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/FilterBookNFTIn"
        - $ref: "#/components/parameters/FilterAccountIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Staking"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /accounts:
    get:
      tags:
        - Account
      summary: Get accounts
      parameters:
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: |-
            The number of items to return.
            When filter.account_in is provided, the limit will be ignored.
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/FilterBookNFTIn"
        - $ref: "#/components/parameters/FilterAccountIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/{evm_address}:
    get:
      tags:
        - Account
      summary: Get an account
      parameters:
        - name: evm_address
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EvmAddress"
        - $ref: "#/components/parameters/FilterBookNFTIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/{evm_address}/stakings:
    get:
      tags:
        - Staking
      summary: Get staking information for an account
      parameters:
        - name: evm_address
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EvmAddress"
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: Optional limit for paginated use cases.
          required: false
          schema:
            type: integer
            nullable: true
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/FilterBookNFTIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Staking"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/{evm_address}/book-nfts:
    get:
      tags:
        - BookNFT
      summary: Get book NFTs for an account
      parameters:
        - name: evm_address
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EvmAddress"
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: Optional limit for paginated use cases.
          required: false
          schema:
            type: integer
            nullable: true
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/AccountBookNFTsRequestSortBy"
        - $ref: "#/components/parameters/AccountBookNFTsRequestSortOrder"
        - $ref: "#/components/parameters/FilterBookNFTIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AccountBookNFT"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/{evm_address}/staking-events/{event_type}:
    get:
      tags:
        - Staking Event
      summary: Get staking events for an account
      parameters:
        - name: evm_address
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EvmAddress"
        - name: event_type
          in: path
          required: true
          schema:
            type: string
            enum:
              - staked
              - unstaked
              - reward-added
              - reward-claimed
              - reward-deposited
              - all
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: Number of items to return.
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/StakingEventsRequestFilterDatetimeFrom"
        - $ref: "#/components/parameters/StakingEventsRequestFilterDatetimeTo"
        - $ref: "#/components/parameters/FilterBookNFTIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StakingEvent"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /book-nfts:
    get:
      tags:
        - BookNFT
      summary: Get book NFTs
      parameters:
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: |-
            The number of items to return.
            When filter.book_nft_in is provided, the limit will be ignored.
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/BookNFTsRequestSortOrderTimeFrame"
        - $ref: "#/components/parameters/BookNFTsRequestTimeFrameSortBy"
        - $ref: "#/components/parameters/BookNFTsRequestTimeFrameSortOrder"
        - $ref: "#/components/parameters/FilterBookNFTIn"
        - $ref: "#/components/parameters/FilterAccountIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookNFT"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /book-nft/{evm_address}:
    get:
      tags:
        - BookNFT
      summary: Get a book NFT
      parameters:
        - name: evm_address
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EvmAddress"
        - $ref: "#/components/parameters/FilterAccountIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookNFT"

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /book-nft/{evm_address}/stakings:
    get:
      tags:
        - Staking
      summary: Get staking information for a book NFT
      parameters:
        - name: evm_address
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EvmAddress"
        - $ref: "#/components/parameters/PaginationRequestKey"
        - name: pagination.limit
          in: query
          description: pagination.limit
          required: false
          schema:
            type: integer
            nullable: true
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/PaginationRequestReverse"
        - $ref: "#/components/parameters/FilterAccountIn"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    $ref: "#/components/schemas/PaginationResponse"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Staking"
                required:
                  - pagination
                  - data

        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    PaginationRequestKey:
      name: pagination.key
      in: query
      description: pagination.key
      required: false
      schema:
        type: integer
    PaginationRequestReverse:
      name: reverse
      in: query
      description: reverse
      required: false
      schema:
        type: boolean
        default: false
...
    AccountBookNFTsRequestSortBy:
      name: sort_by
      in: query
      required: false
      schema:
        type: string
        enum:
          - pool_share
          - staked_amount
          - pending_reward_amount
        default: staked_amount
    AccountBookNFTsRequestSortOrder:
      name: sort_order
      in: query
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
    StakingEventsRequestFilterDatetimeFrom:
      name: filter_datetime_from
      in: query
      required: false
      schema:
        type: string
        format: date-time
    StakingEventsRequestFilterDatetimeTo:
      name: filter_datetime_to
      in: query
      required: false
      schema:
        type: string
        format: date-time
    FilterBookNFTIn:
      name: filter_book_nft_in
      in: query
      required: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/EvmAddress"
    FilterAccountIn:
      name: filter_account_in
      in: query
      required: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/EvmAddress"
...
  schemas:
    Staking:
      type: object
      properties:
        book_nft:
          $ref: "#/components/schemas/EvmAddress"
        account:
          $ref: "#/components/schemas/EvmAddress"
        pool_share:
          $ref: "#/components/schemas/Percentage"
        staked_amount:
          $ref: "#/components/schemas/Uint256"
        pending_reward_amount:
          $ref: "#/components/schemas/Uint256"
        claimed_reward_amount:
          $ref: "#/components/schemas/Uint256"
      required:
        - book_nft
        - account
        - pool_share
        - staked_amount
        - pending_reward_amount
        - claimed_reward_amount
...
    Account:
      type: object
      properties:
        evm_address:
          $ref: "#/components/schemas/EvmAddress"
        staked_amount:
          $ref: "#/components/schemas/Uint256"
        pending_reward_amount:
          $ref: "#/components/schemas/Uint256"
        claimed_reward_amount:
          $ref: "#/components/schemas/Uint256"
      required:
        - evm_address
        - staked_amount
        - pending_reward_amount
        - claimed_reward_amount
...
    PaginationResponse:
      type: object
      properties:
        next_key:
          type: integer
        count:
          type: integer
      required:
        - next_key
        - count
    EvmAddress:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
      example: "0x1234567890123456789012345678901234567890"
...
    Percentage:
      type: string
      pattern: "^[0-9](\\.[0-9]+)?$"
      example: "0.1234567890"
    Error:
      description: Represents error object
      type: object
      properties:
        code:
          type: integer
          format: int64
          default: 500
        message:
          type: string
        sentry_error_id:
          type: string
      required:
        - code
        - message